decorate rails in scripts
